Refactored models & endpoints

+++ backend/app/models.py
@@ -13,16 +13,15 @@ FOLLOWUPS = [
     for days in [0, 3, 5, 7, 14, 30]
 ]
 
-
+# TODO: explain tear down
 def initialize_db(app, tearDown=False):
     app.app_context().push()
     db.init_app(app)
 
-    if tearDown:
+    # It be really bad if teardown was true on production!
+    if tearDown and app.config["DEBUG"]:
         db.drop_all()
         db.create_all()
-
-
 
 class ConstantsModel(db.Model):
     __tablename__ = "constants"
@@ -47,6 +46,10 @@ class TimestampModel:
         db.DateTime, onupdate=datetime.utcnow, default=datetime.utcnow
     )
 
+lead_user = db.Table('lead_user', db.Model.metadata,\
+                    db.Column('leads', db.Integer, db.ForeignKey('leads.id')), \
+                    db.Column('users', db.Integer, db.ForeignKey('users.id'))
+                    )
 
 class Lead(TimestampModel, db.Model):
     __tablename__ = "leads"
@@ -67,6 +70,9 @@ class Lead(TimestampModel, db.Model):
 
     notInterested = db.Column(db.Boolean, default=False)
     hasApplied = db.Column(db.Boolean, default=False)
+    hasLeased = db.Column(db.Boolean, default=False)
+
+    users = db.relationship("User", secondary=lead_user, backref="leads")
 
     def __init__(self):
         self.followups = ConstantsModel.query.first().followupTemplate
@@ -74,16 +80,17 @@ class Lead(TimestampModel, db.Model):
     def __repr__(self):
         return f"<Lead {self.id}>"
 
-# Apartments/user models --remove before pushing
-
 class Apartment(db.Model):
     __tablename__ = "apartments"
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    name = db.Column(db.String(200), default="")
+    aptName = db.Column(db.String(200), default="")
+    website = db.Column(db.String(200), default="")
+    units = db.Column(db.Integer, default=0)
+    propertyType = db.Column(db.String(200), default="")
+    websiteType = db.Column(db.String(200), default="")
 
     def __repr__(self):
         return f"<Apartment {self.id}>"
-
 
 # https://stackoverflow.com/questions/18807322/sqlalchemy-foreign-key-relationship-attributes
 class User(db.Model):
@@ -93,11 +100,9 @@ class User(db.Model):
     phoneNumber = db.Column(db.String(50), default="")
     email = db.Column(db.String(200), default="")
     role = db.Column(db.String(50), default="") 
-    apartment = db.Column(db.Integer, db.ForeignKey(Apartment.id))
+    apartment_id = db.Column(db.Integer, db.ForeignKey(Apartment.id))
+
+    apartment = db.relationship("Apartment")
 
     def __repr__(self):
         return f"<User {self.id}>"
-
-
-
-
